+++ docs/source/conf.py
@@ -67,6 +67,7 @@ intersphinx_mapping = {
 
 # By default, sort documented members by type within classes and modules.
 autodoc_member_order = "bysource"
+autodoc_default_options = {"show-inheritance": True, "undoc-members": True}
 
 # Include default values when documenting parameter types.
 typehints_defaults = "comma"